% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesCombo.R
\docType{data}
\name{BayesCombo}
\alias{BayesCombo}
\title{BayesCombo data set}
\format{A data frame with 15124 rows of gene probes and 18 columns.
\describe{
  \item{beta.sk:}{The effect size from the gene expression output for the skeletal muscle tissue.}
  \item{beta.ab:}{The effect size from the gene expression output for the abdominal fat tissue.}
  \item{se.sk:}{The standard error from the gene expression output for the skeletal muscle tissue.}
  \item{se.ab:}{The standard error from the gene expression output for the abdominal fat tissue.}
  \item{Aexprs:}{Two columns of of the Average expression of the two tissue.}
  \item{rawdata:}{Six columns of log2 scale raw data from the skeletal muscle study
  and six columns of log2 scale raw data from the abdominal fat study.} }}
\description{
Subset of 'Transforming growth factor beta-like stimulated
clone 22 D4 promotes diabetic hyperglycemia and insulin resistance
skeletal muscle and abdominal muscle'.
}
\details{
The study design has 4 sets of 7 mice, a sham group ,
a TSC22D4-knockout group, a LCN13-knockout and a TSC22D4-knockout LCN13-knockout group.
Only the sham and the TSC22D4-knockout group are used in this data set. Due to file size
60 % of rows are randomly selected from the original dataset.
}
\references{
Transforming growth factor beta-like stimulated clone 22 D4
 promotes diabetic hyperglycemia and insulin resistance.
 Friedrich K, Jones A, Seibert O, Sijmonsma TP, Wang X, Sticht C, Gretz N, Fleming T,
  Nawroth PP, Stremmel W, Rose AJ, Diaz MB, Bluher M, Herzig S
}

