% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSfactor.R
\name{BSfactor}
\alias{BSfactor}
\title{Calculate Bayes Safety factor}
\usage{
BSfactor(beta1, var1, threshold = 0.95, beta0 = 0, n = 100,
  hypothesis = 1, reverse = FALSE)
}
\arguments{
\item{beta1}{Vector of observed means}

\item{var1}{Vector of observed variances}

\item{threshold}{Sets the value at which the border is set}

\item{beta0}{Vector or value of prior means. Default is  0}

\item{n}{Number of steps between 0.33333 and 0.99999}

\item{hypothesis}{Parameter has different functionalities depending on value of reverse}

\item{reverse}{Changes the hypothesis in which the prior model probability increases. If this is TRUE then the value of the hypothesis parameter will determine the
hypothesis which will increase in probability}
}
\value{
Object of class "BSfactor" which contains a matrix of final posterior model probabilities, a matrix of prior model probabilities,
the 'boundary' and which hypothesis is being considered
}
\description{
Calculates how much can the prior model probability be weighted to the null hypothesis.
}
\details{
Incrementally increases a the prior posterior probability for a given hypothesis, and shows the resulting posterior
model probabilities
}
\examples{
x <- BSfactor( beta1 = c(0.068,-0.084,0.175,0.337),
     var1 = c(0.000841,0.002916,0.008649,0.032041),
     beta0 = 0 )
}
\seealso{
\code{\link{plot.BSfactor}, \link{seq}}
}

