% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc.post.beta}
\alias{calc.post.beta}
\alias{calc.post.se}
\alias{prior.se}
\title{Utility functions}
\usage{
calc.post.beta(beta, se.beta, beta0, se0)

calc.post.se(se.beta, se0)

prior.se(beta, se.beta, percent)
}
\arguments{
\item{beta}{An numeric effect size.}

\item{se.beta}{A numeric standard error of the effect.}

\item{beta0}{A numeric prior value for the effect size.}

\item{se0}{A numeric prior variance.}

\item{percent}{A numeric value that is used to calculate the prior
variance.}
}
\description{
Not called directly by user.
}

