% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.PMP.R
\name{calculate.PMP}
\alias{calculate.PMP}
\title{Calculate updated Posterior Model Proababilities}
\usage{
calculate.PMP(beta1, var1, pi0, beta0, n = 10000, var.mult = 1)
}
\arguments{
\item{beta1}{Vector of observed means}

\item{var1}{Vector of observed variances}

\item{pi0}{Vector of prior model probabilities}

\item{beta0}{Vector or value of prior means (usually 0)}

\item{n}{Number of times distributions are sampled}

\item{var.mult}{Multiplier of the prior variances}
}
\value{
Object of class BFcombo which contains a matrix of posterior model probabilities for each updated step.
}
\description{
Wrapper function to caclulate the posterior model probabilities from the observed means and variances.
}
\details{
Creates the BFcombo object from the inputs and runs a pipeline to produce the complete set of posterior model
probabilities given the inputs.
}
\examples{
x <- calculate.PMP( beta1 = c(0.090,0.140,1.090,1.781),
       var1 = c(0.000841,0.002916,0.008649,0.032041),
       beta0 = 0, pi0 = rep(1/3,3))
}
\seealso{
\code{\link{prior.var} \link{u.post.param} \link{unconstrained.BF} \link{PMP.update}}
}

