% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.BFcombo.R
\name{make.BFcombo}
\alias{make.BFcombo}
\title{Create object of class 'BFcombo'}
\usage{
make.BFcombo(beta1, var1, beta0, pi0)
}
\arguments{
\item{beta1}{Vector of observed means}

\item{var1}{Vector of observed variances}

\item{beta0}{Value of prior mean (usually 0)}

\item{pi0}{Vector of prior model probabilities}
}
\description{
Many of the BayesCombo functions require objects of the class 'BFcombo' to run.
}
\details{
This function is called within the prior.var function.
}
\seealso{
\code{\link{prior.var}}
}

