% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi0.to.1.R
\name{pi0.to.1}
\alias{pi0.to.1}
\title{Calculate the posterior model probabilities given the prior model probabilities
and the unconstrained Bayes factors.}
\usage{
pi0.to.1(pi0, BF)
}
\arguments{
\item{pi0}{Vector of prior model probabilities}

\item{BF}{Vector of Bayes factors which correspond to the prior model probabilities}
}
\description{
Calculates the posterior model probabilities given the prior model probabilities
and the unconstrained Bayes factors, this is then updated by setting the posterior model probabiities to the prior
model probabilities
}
\details{
This function is called from within the update.PMP function
}
\seealso{
\code{\link{PMP.update}}
}

