% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BFcombo.R
\name{plot.BFcombo}
\alias{plot.BFcombo}
\title{Plots effect sizes, confidence intervals, and prior variances}
\usage{
\method{plot}{BFcombo}(x, ...)
}
\arguments{
\item{x}{A object of the class BFcombo (see make_BFcombo)}

\item{...}{additional plotting parameters.}
}
\value{
Plot of prior and observed confidence intervals.
}
\description{
Produces a visualisation of prior variances by plotting the
confidence intervals around the prior effect size and also shows the observed
effect size confidence intervals.
}
\examples{
x <- calculate.PMP( beta = c(0.0126, 5.0052, 1.2976, 0.0005),
       se.beta = c(0.050, 2.581, 2.054, 0.003) )
plot(x)
}
\seealso{
\code{\link{calculate.PMP}}
}

