% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PMPlist.R
\name{plot.PMPlist}
\alias{plot.PMPlist}
\title{Plot of posterior model probabilities}
\usage{
\method{plot}{PMPlist}(x, leg.loc = "topleft", lty = 1, ...)
}
\arguments{
\item{x}{An object of the class \code{PMPlist} created by the
\code{pmp.combo()} function.}

\item{leg.loc}{Location of the legend; default is top left. See the
\code{legend()} help pages for all the options.}

\item{lty}{An vecter of integers that specifies the line types. See the
\code{matplot()} function for details.}

\item{...}{Other options passed to \code{matplot()}.}
}
\value{
Plot of posterior model probabilities.
}
\description{
Shows how the posterior model probabilities evolve as each study
is added to the analysis.
}
\details{
Each line in the graph is a hypothesis and the lines start at the
prior model probabilities (usually 1/3).
}
\examples{
x <- pmp.combo(beta = c(0.0126, 5.0052, 1.2976, 0.0005),
       se.beta = c(0.050, 2.581, 2.054, 0.003) )
plot(x)
}
\seealso{
\code{\link{pmp.combo}}
}

