% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMP.plot.R
\name{PMP.plot}
\alias{PMP.plot}
\title{Plot of Posterior model probabilities}
\usage{
PMP.plot(BFcombo, pi0 = rep(1/3, 3))
}
\arguments{
\item{BFcombo}{A object of the class BFcombo which contains the PMP section}

\item{pi0}{Vector of prior model probabilities}
}
\value{
Plot of posterior model probabilities.
}
\description{
Produces a visualisation of posteriror model probabilities by plotting the probabilities at each updating step.
}
\details{
There is a line for each hypothesis and the starting point is the prior model probabilities for each
hypothesis.
}
\examples{
x <- calculate.PMP(beta1 = c(0.0126474408, 5.0051724138, 1.2975612498, 0.0004762455),
       var1 = c(2.538974e-03, 6.662216e+00, 4.219142e+00, 6.963380e-06),
    beta0 = 0, pi0 = rep(1/3,3) )
PMP.plot(x)
}
\seealso{
\code{\link{calculate.PMP}}
}

