% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmp.update.R
\name{pmp.update}
\alias{pmp.update}
\title{Calculates posterior model probabilities for multiple studies}
\usage{
pmp.update(beta, se.beta, beta0 = 0, percent = 99, H0 = c(0, 0),
  scale = FALSE, mod.priors = rep(1/3, 3), var.mult = 1, adjust = FALSE,
  epsilon = 1e-06, adj.factor = 1e-04, ...)
}
\arguments{
\item{beta}{An numeric effect size.}

\item{se.beta}{A numeric standard error of the effect.}

\item{beta0}{A numeric prior value for the effect size. Default is zero.}

\item{percent}{A numeric value that is used to calculate the prior
variance. The default value of 99 calculates the prior variance so that the
99% confidence intervals of the prior are aligned with the largest (furthest
from zero) confidence intervals of the data distribution.}

\item{H0}{A vector of length two that defines the null hypothesis. If the
values are identical (e.g. \code{H0 = c(0,0)}) a point null is used,
otherwise the null is defined as the range between the lower and upper value.}

\item{scale}{Logical. Whether to scale the effect size by its standard
error. This has no effect on the calculations but it may be useful to use
standardised effect sizes.}

\item{mod.priors}{Prior model probabilities; default is an equal probability
of 1/3.}

\item{var.mult}{Variance multiplier used to increase or decrease the prior
variance.}

\item{adjust}{Logical, whether to adjust the posterior model probabilities if
they are too small. When combining multiple studies, this prevents a single
study from having too much influence on the results. For example, if the
probability for a hypothesis is zero, then additional studies cannot alter
this probability.}

\item{epsilon}{A small value that a posterior model probability must be below
before an adjustment is made.}

\item{adj.factor}{A small number added to each posterior model probability if
\code{adjust=TRUE} and a PMP is less than \code{epsilon}. The PMPs are then
re-scaled to sum to one.}

\item{...}{Options to be passed to \code{\link{pmp}}.}
}
\value{
Object of class \code{PMPlist} which contains a matrix of posterior
model probabilities for each updated step and other calculate values
}
\description{
The function takes multiple effect size and a their standard
errors and calculates the posterior model probabilities for each hypothesis
(H<: the effect size is less than 0, H0: the effect size is zero, or H>: the
effect size is greater than zero).
}
\details{
This function calls \code{\link{pmp}} once for each study to be
combined where the posterior model probabilities for one study are used as
the priors for the next study. The inputs are identical to the
\code{\link{pmp}} function, so please see that documentation for details.
}
\examples{
x <- pmp.update(beta = c(0.0126, 5.0052, 1.2976, 0.0005),
       se.beta = c(0.050, 2.581, 2.054, 0.003) )
x
plot(x)
}

