% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PMP}
\alias{print.PMP}
\alias{print.PMPlist}
\alias{summary.BSfactor}
\alias{summary.PMP}
\alias{summary.PMPlist}
\title{Print and summary methods for PMP, PMPlist, and BSfactor objects}
\usage{
\method{summary}{PMP}(x, digits = 4)

\method{print}{PMP}(x, digits = 4)

\method{summary}{PMPlist}(x, digits = 4)

\method{print}{PMPlist}(x, digits = 4)

\method{summary}{BSfactor}(x, digits = 4)
}
\arguments{
\item{x}{A \code{PMP}, \code{PMPlist}, or \code{BSfactor} object.}

\item{digits}{Number of digits to show.}
}
\value{
A list of posterior model probabilities and other calculated values.
}
\description{
Print and summary methods for PMP, PMPlist, BSfactor objects.
}

