% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PPH}
\alias{print.EV}
\alias{print.PPH}
\alias{summary.EV}
\alias{summary.PPH}
\title{Print and summary methods for PPH and EV objects}
\usage{
\method{summary}{PPH}(x, digits = 4)

\method{print}{PPH}(x, digits = 4)

\method{summary}{EV}(x, digits = 4)

\method{print}{EV}(x, digits = 4)
}
\arguments{
\item{x}{A \code{PPH} or \code{EV} object.}

\item{digits}{Number of digits to show.}
}
\value{
A list of posterior hypothesis probabilities and other calculated
values.
}
\description{
Print and summary methods for PPH and EV objects.
}

