% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSfactor.R
\name{BSfactor}
\alias{BSfactor}
\title{Calculate Bayes Safety factor}
\usage{
BSfactor(beta1, var1 = NULL, beta0 = 0, n = 100, threshold = 0.95,
  hypothesis = 1, reverse = FALSE)
}
\arguments{
\item{beta1}{Vector of observed means or an object of class 'BFcombo'}

\item{var1}{Vector of observed variances or NULL if beta1 is class 'BFcombo'}

\item{beta0}{Vector or value of prior means. Default is  0}

\item{n}{Number of steps between 0.33333 and 0.99999}

\item{threshold}{Sets the value at which the border is set}

\item{hypothesis}{Parameter has different functionalities depending on value of reverse}

\item{reverse}{Changes the hypothesis in which the prior model probability increases. If this is TRUE then the value of the hypothesis parameter will determine the
hypothesis which will increase in probability}
}
\value{
Object of class 'BSfactor' which contains a matrix of final posterior model probabilities, a matrix of prior model probabilities,
the 'boundary' and which hypothesis is being considered
}
\description{
Calculates how much the prior model probability be weighted to a given hypothesis untill the result of running calculate.PMP will change for a given threshold.
}
\details{
Incrementally increases a the prior posterior probability for a given hypothesis, and shows the resulting posterior
model probabilities
}
\examples{

x <- BSfactor( beta1 = c(0.0126474408, 5.0051724138, 1.2975612498, 0.0004762455),
       var1 = c(2.538974e-03, 6.662216e+00, 4.219142e+00, 6.963380e-06),
       beta0 = 0, reverse = TRUE )

}
\seealso{
\code{\link{plot.BSfactor}, \link{seq}}
}

