% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc.post.beta}
\alias{calc.post.beta}
\alias{calc.post.se}
\alias{prior.se}
\title{Utility functions}
\usage{
calc.post.beta(beta, se.beta, beta0, se0)

calc.post.se(se.beta, se0)

prior.se(beta, se.beta, ci)
}
\arguments{
\item{beta}{Effect size.}

\item{se.beta}{Standard error of the effect.}

\item{beta0}{Prior for the effect size.}

\item{se0}{Prior standard error.}

\item{ci}{A numeric value that is used to calculate the prior standard error.}
}
\description{
Not called directly by user.
}

