% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BFcombo.R
\name{plot.BFcombo}
\alias{plot.BFcombo}
\title{Plot of Prior Variances given Observed Means and Variances}
\usage{
\method{plot}{BFcombo}(BFcombo, percent = 99)
}
\arguments{
\item{BFcombo}{A object of the class BFcombo (see make_BFcombo).}

\item{percent}{Sets the percentage of the confidence interval. Default is 99.}
}
\value{
Plot of Prior and observed confidence intervals.
}
\description{
Produces a visualisation of prior variances by plotting the confidence intervals around the
 prior mean and also shows the observed means and variances as confidence intervals.
}
\examples{
x <- calculate.PMP( beta1 = c(0.090,0.140,1.090,1.781),
       var1 = c(0.000841,0.002916,0.008649,0.032041),
       beta0 = 0, pi0 = rep(1/3,3) )
plot(x)
}
\seealso{
\code{\link{calculate.PMP}}
}

