% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BSfactor.R
\name{plot.BSfactor}
\alias{plot.BSfactor}
\title{Plots Bayesian safety (BS) factor results}
\usage{
\method{plot}{BSfactor}(x, ...)
}
\arguments{
\item{x}{object of the class BSfactor}

\item{...}{additional parameters not required}
}
\value{
Plot of posterior model probabilities in relation to changes in the
prior model probabilities.
}
\description{
Produces a visualisation of the Bayes safety factor results.
}
\details{
Each solid line is a hypothesis, the x-axis is the prior model
probability for the null hypothesis and the y-axis is the posterior model
probability. The dotted line is the threshold.
}
\examples{
x <- BSfactor(beta = c(0.0126, 5.005, 1.298, 0.000476),
       se.beta = c(0.0504, 2.5811, 2.0541, 0.0026),
       beta0 = 0, reverse = TRUE )
plot(x)
}
\seealso{
\code{\link{BSfactor}}
}

