% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BSfactor.R
\name{plot.BSfactor}
\alias{plot.BSfactor}
\title{Plot Bayes Safety factor results}
\usage{
\method{plot}{BSfactor}(x, ...)
}
\arguments{
\item{x}{object of the class BSfactor}

\item{...}{additional parameters not required}
}
\value{
Plot of posterior model probabilities in relation to changes in the prior model probabilities.
}
\description{
Produces a visualisation of the Bayes safety factor results.
}
\details{
Each solid line is a hypothesis, the x-axis is of the prior model proability for the null hypothesis,
the y-axis is the posterior model probability. The dotted line is the threshold.
}
\examples{
x <- BSfactor( beta1 = c(0.068,-0.084,0.175,0.337),
       var1 = c(0.000841,0.002916,0.008649,0.032041),
       beta0 = 0 )
plot(x)
}
\seealso{
\code{\link{BSfactor}}
}

