% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMP.plot.R
\name{PMP.plot}
\alias{PMP.plot}
\title{Plot of Posterior model probabilities}
\usage{
PMP.plot(BFcombo, pi0 = rep(1/3, 3))
}
\arguments{
\item{BFcombo}{A object of the class BFcombo which contains the PMP section}

\item{pi0}{Vector of prior model probabilities}
}
\value{
Plot of posterior model probabilities.
}
\description{
Produces a visualisation of posteriror model probabilities by plotting the probabilities at each updating step.
}
\details{
There is a line for each hypothesis and the starting point is the prior model probabilities for each
hypothesis.
}
\examples{
x <- calculate.PMP( beta1 = c(0.090,0.140,1.090,1.781),
       var1 = c(0.000841,0.002916,0.008649,0.032041),
       beta0 = 0, pi0 = rep(1/3,3) )
PMP.plot(x)
}
\seealso{
\code{\link{calculate.PMP}}
}

