% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMP.plot.R
\name{PMP.plot}
\alias{PMP.plot}
\title{Plot of posterior model probabilities}
\usage{
PMP.plot(BFcombo, pi0 = rep(1/3, 3))
}
\arguments{
\item{BFcombo}{A object of the class BFcombo which contains the PMP section.}

\item{pi0}{Vector of prior model probabilities.}
}
\value{
Plot of posterior model probabilities.
}
\description{
Plots the updated posterior model probabilities as each study
is added to the total.
}
\details{
Each line is a hypothesis and they start at the prior model
probabilities (usually 1/3).
}
\examples{
x <- calculate.PMP( beta = c(0.0126, 5.0052, 1.2976, 0.0005),
       se.beta = c(0.050, 2.581, 2.054, 0.003) )
PMP.plot(x)
}
\seealso{
\code{\link{calculate.PMP}}
}

