% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.var.R
\name{prior.var}
\alias{prior.var}
\title{Calculate the prior variances for the unconstrained model}
\usage{
prior.var(beta1, var1, beta0, pi0, percent = 99)
}
\arguments{
\item{beta1}{Vector of observed means}

\item{var1}{Vector of observed variances}

\item{beta0}{Vector or value of prior means (usually 0)}

\item{pi0}{Vector of prior model probabilities}

\item{percent}{Confidence interval precentage}
}
\value{
vector of equal length to that of the observed variances.
}
\description{
Given a set of values for the observed mean,variance and the prior mean, a set of unconstrained prior variances are calculated.
}
\details{
The prior variances are calculated by getting the largest confidence interval from the prior mean by using the
confidence interval of the observed values.
}

