% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BFcombo.R
\name{summary.BFcombo}
\alias{summary.BFcombo}
\title{Summarise a BFcombo object}
\usage{
\method{summary}{BFcombo}(object, ...)
}
\arguments{
\item{object}{A object of the class BFcombo}

\item{...}{Additional parameters}
}
\value{
A summary of the BFcombo object passed.
}
\description{
Gives summary of various sections of the BFcombo class.
}
\details{
For all the sections which provide a mean or variance vector, the
summary function will return a minimum and maximum value.  For a Bayes factor
or probability, the summary function will return a matrix with the values
rounded to 4 decimal points.
}
\examples{
x <- calculate.PMP( beta = c(0.0126, 5.0052, 1.2976, 0.0005),
       se.beta = c(0.050, 2.581, 2.054, 0.003) )
summary(x)
}
\seealso{
\code{ \link{calculate.PMP} \link{prior.var}}
}

