% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BFcombo.R
\name{summary.BFcombo}
\alias{summary.BFcombo}
\title{Summarize a BFcombo object}
\usage{
\method{summary}{BFcombo}(object, ...)
}
\arguments{
\item{object}{A object of the class BFcombo}

\item{...}{Additional parameters}
}
\value{
A summary of the BFcombo object passed.
}
\description{
Gives summary of various sections of the BFcombo class.
}
\details{
For all the sections which provide a mean or variance vector, the summary function will return a minimum and maximim value.
For a Bayes factor or probability, the summary function will return a matrix with the values rounded to 4 decimal points.
}
\examples{
x <- calculate.PMP( beta1 = c(0.090,0.140,1.090,1.781),
     var1 = c(0.000841,0.002916,0.008649,0.032041),
     beta0 = 0, pi0 = rep(1/3,3))
summary(x)
}
\seealso{
\code{ \link{calculate.PMP} \link{prior.var}}
}

