% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BSfactor.R
\name{summary.BSfactor}
\alias{summary.BSfactor}
\title{Summarise a BSfactor object}
\usage{
\method{summary}{BSfactor}(object, ...)
}
\arguments{
\item{object}{A object of the class BSfactor}

\item{...}{Additional parameters not required}
}
\value{
A summary of the BSfactor object.
}
\description{
Gives summary of various sections of the BSfactor class.
}
\details{
For all section of the BSfactor object which are matrices, the
summary function will return the first and last rows with the values rounded
to 4 decimal points.  The summary also contains other key pieces of
information such as boundary and threshold.
}
\examples{
x <- BSfactor( beta = c(0.0126, 5.0052, 1.2976, 0.0005),
       se.beta = c(0.050, 2.581, 2.054, 0.003),
       beta0 = 0, reverse = TRUE )
summary(x)
}
\seealso{
\code{ \link{BSfactor}}
}

