% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.post.params.R
\name{u.post.param}
\alias{u.post.param}
\title{Calculate unconstrained Posterior Model parameters.}
\usage{
u.post.param(BFcombo)
}
\arguments{
\item{BFcombo}{A object of the class BFcombo but must also incldes a section with unconstrained prior variances}
}
\value{
A set of unconstrained posterior variances and means.
}
\description{
Calculates the unconstrained posterior model parameters which will then be used to calculate
unconstained Bayes factors.
}
\details{
See vignette for detailed explanation.
}
\examples{
x <- prior.var(beta1 = c(0.090,0.140,1.090,1.781),
     var1 = c(0.000841,0.002916,0.008649,0.032041),
     beta0 = 0, pi0 = rep(1/3,3) )
x <- u.post.param(x)
}
\seealso{
\code{\link{prior.var}}
}

